requirements.txt
fastapi>=0.104.0,<1.0
uvicorn>=0.24.0,<1.0
streamlit>=1.28.0,<2.0
polars>=0.19.0,<2.0
pandas>=2.1.0,<3.0
duckdb>=1.1.0,<2.0
plotly>=5.18.0,<6.0
matplotlib>=3.8.0,<4.0
seaborn>=0.13.0,<1.0
scikit-learn>=1.3.0,<2.0
langchain>=0.1.0,<0.2
langgraph>=0.0.20,<1.0
chromadb>=0.4.18,<0.5
sentence-transformers>=2.2.0,<3.0
ollama>=0.1.0,<1.0
python-multipart>=0.0.6,<1.0
reportlab>=4.0.0,<5.0
openpyxl>=3.1.0,<4.0
python-dotenv>=1.0.0,<2.0
scipy>=1.11.0,<2.0
numpy>=1.26.0,<3.0
requests>=2.31.0,<3.0
kaleido>=0.2.1,<2.0
groq>=0.4.0,<2.0
pytest>=7.4.0,<9.0
